+++ Sources/CodexSwitchboard/CodexAccountCaptureService.swift
@@ -1273,7 +1273,7 @@ private extension Data {
     }
 }
 
-private extension ISO8601DateFormatter {
+extension ISO8601DateFormatter {
     static let codexSwitchboard: ISO8601DateFormatter = {
         let formatter = ISO8601DateFormatter()
         formatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]